refactor(landing): split App.tsx (2225 -> 47 l.)
App.tsx now only composes the root layout. Everything else moves out:

- components/buttons/SceneBtn.tsx (ex-DeckBtn, renamed for SceneStack
  parity) + components/buttons/PrimaryBtn.tsx
- components/cards/{OpenCard,ProductCard,ValueCard}.tsx — ValueCard
  cross-imports App.module.css to keep the .placeholder a:hover
  cascade on .valuePill
- scenes/zones.tsx — ZONE_NODES + ZONE_LABELS + PARTNERS
- scenes/PlaceholderSlide.tsx — PlaceholderSlide + SlideLayout +
  LAYOUT_AREAS + layoutClass + PlaceholderSlideProps
- scenes/slides.config.tsx — SLIDES, SUB_STATES, VARIANTS, REVEAL_BGS,
  SlideConfig (uses .tsx because headerContent carries inline JSX)
